Refactoring
Rename filenames to use similar convention
Split components up for reuse
Create new component folders
Simplify operation specification (custom vs all)
Split settings page into 2 new pages - backup and dapps
