+++ src/components/qr/Scan.vue
@@ -4,13 +4,13 @@
     import { QrcodeStream } from 'qrcode-reader-vue3'
 
     import { ipcRenderer } from "electron";
-    import store from '../store/index';
+    import store from '../../store/index';
 
     import {
         injectedCall,
         voteFor,
         transfer
-    } from '../lib/apiUtils.js';
+    } from '../../lib/apiUtils.js';
 
     const { t } = useI18n({ useScope: 'global' });
     let qrInProgress = ref(false);
