+++ src/router/index.js
@@ -6,7 +6,8 @@ import Start from "../components/start";
 import Dashboard from "../components/dashboard";
 import Restore from "../components/restore";
 import AddAccount from "../components/add-account";
-import Settings from "../components/settings";
+import Dapps from "../components/dapps";
+import Backup from "../components/backup";
 import Totp from "../components/totp";
 import Qr from "../components/qr";
 import Popups from "../components/popups";
@@ -20,9 +21,16 @@ const router = createRouter({
       }
     },
     {
-      path: '/settings',
+      path: '/backup',
       components: {
-        default: Settings,
+        default: Backup,
+        header: HeaderSmall
+      }
+    },
+    {
+      path: '/dapps',
+      components: {
+        default: Dapps,
         header: HeaderSmall
       }
     },
